change: replaced all pristine gem textures (cred: rappenem) change: added pristine amethyst, found in tetra geodes tech: renamed all aspects from camelCase to snake_case tech: moved chthonic extractor hammering config option to feature flag change: reworked workbench visuals change: added aspect info to module details view tech: put multislot schematics in a working state fix: fixed sweeping strike not alternating direction without truesweep change: changed sweeping strike drain: now drains 0.5 at a minimum rather than adding 0.5 to drain for each block fix: fixed how reaching affects sweeping strike to align with tooltip (reduced the entire drain, not just drain from block hardness) tech: moved some localization entries tech: fixed multislot modules defaulting texture namespaces back to tetra tech: fixed source information merging in the wrong direction tech: fixed schematic module requirement not matching the material field fix: added expanded tooltips for medial, lateral and axial limiters to explain what that actually means change: efficiency honing for stonecutters to start out stronger (still hits 15 at rank V) change: the cobblestone material should now better match modded cobblestone tech: fixed long and recurve bow not picking up improvements from their respective improvement folder fix: fixed spearheads accepting tool enchantments fix: fixed looted stonecutter not being repairable with scrap

+++ src/main/java/se/mickelus/tetra/gui/ZOffsetGui.java
@@ -4,9 +4,9 @@ import net.minecraft.client.gui.GuiGraphics;
 import se.mickelus.mutil.gui.GuiElement;
 
 public class ZOffsetGui extends GuiElement {
-    protected int z;
+    protected double z;
 
-    public ZOffsetGui(int x, int y, int z) {
+    public ZOffsetGui(int x, int y, double z) {
         super(x, y, 0, 0);
         this.z = z;
     }
